chore(ci): update build image (#29)
A client has a GNU C incompatibility when using auto injection for RUM
in RHEL8.8.

We provided a fix in nginx which should also fixes this issue. The fix
is in the last version of ngx_musl_toolchain so we update the httpd ci
image to benefit from the fix